Integrasi CRUD [Organiasi, Wil Rw, Wil Rt

+++ src/pages/organisasi/Organisasi.js
@@ -5,6 +5,7 @@ import axios from "axios";
 import { api_organisasi } from "../../utils/api";
 import Footer from "../../component/Footer";
 import Sidebar from "../../component/Sidebar";
+import Swal from "sweetalert2";
 
 const authConfig = {
   headers: {
@@ -17,6 +18,10 @@ function Organisasi() {
   const [ModalEditOpen, setModalEditOpen] = useState(false);
   const [datas, setDatas] = useState([]);
   const [nama_organisasi, setNama_organisasi] = useState("");
+  const [pages, setPages] = useState(0);
+  const [show, setShow] = useState(false);
+  const [currentPage, setCurrentPage] = useState(0);
+  const size = 5;
   const [id, setId] = useState(0);
   const navigate = useNavigate();
 
@@ -38,14 +43,28 @@ function Organisasi() {
     setModalEditOpen(false);
   };
 
+
+  const handlePageChange = (newPage) => {
+    if (newPage < 0 || newPage > pages) {
+      return; // Jangan lakukan apa pun jika halaman baru di luar rentang yang valid
+    }
+    setCurrentPage(newPage);
+    getOrganisasi(newPage);
+  };
+
   const getOrganisasi = async () => {
     try {
-      const response = await axios.get(`${api_organisasi}/all`, authConfig);
-      setDatas(response.data.data);
+      const response = await axios.get(
+        `http://localhost:8000/e-kampoeng/api/organisasi?page=${pages}&size=${size}`,
+        authConfig
+      );
+      setPages(response.data.data.totalPages);
+      setDatas(response.data.data.content);
     } catch (error) {
-      console.log(error);
+      alert("Terjadi Kesalahan: " + error);
     }
   };
+
 
   const getOrganisasiById = async (id) => {
     try {
@@ -56,14 +75,28 @@ function Organisasi() {
     }
   };
 
+
+
   const addOrganisasi = async (e) => {
     e.preventDefault();
+    e.persist();
+
     const req = {
       nama_organisasi: nama_organisasi,
     };
     try {
-      await axios.post(`${api_organisasi}/add`, req, authConfig);
-      window.location.reload();
+      await axios.post(api_organisasi, req, authConfig);
+      setShow(false);
+      Swal.fire({
+        icon: "success",
+        title: "Sukses Menambahkan",
+        showConfirmButton: false,
+        timer: 2500,
+      });
+      setTimeout(() => {
+        navigate("/organisasi");
+        window.location.reload();
+      }, 2500);
     } catch (error) {
       console.log(error);
     }
@@ -77,26 +110,55 @@ function Organisasi() {
     try {
       await axios.put(`${api_organisasi}/${id}`, req, authConfig);
       setModalEditOpen(false);
-      window.location.reload();
+      Swal.fire({
+        icon: "success",
+        showConfirmButton: false,
+        timer: 1500,
+        title: "Berhasil Mengedit",
+      }).then(() => {
+        window.location.reload();
+      });
     } catch (error) {
-      console.log(error);
+      console.error("Terjadi kesalahan:", error);
     }
   };
 
   const deleteOrganisasi = async (ID) => {
     try {
-      await axios.delete(`${api_organisasi}/${ID}`, authConfig);
-      setModalEditOpen(false);
-      window.location.reload();
+      await Swal.fire({
+        title: "Apakah Anda yakin?",
+        text: "Anda tidak akan dapat mengembalikan ini!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Ya, hapus!",
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          await axios.delete(`${api_organisasi}/${ID}`, authConfig);
+          Swal.fire({
+            title: "Terhapus!",
+            text: "Data telah dihapus.",
+            icon: "success",
+            timer: 1500,
+            showConfirmButton: false,
+          }).then(() => {
+            window.location.reload();
+          });
+        }
+      });
     } catch (error) {
-      console.log(error);
+      console.error("Terjadi kesalahan:", error);
+      Swal.fire({
+        icon: "error",
+        title: "Gagal",
+        text: "Terjadi kesalahan saat menghapus data.",
+      });
     }
   };
-
   useEffect(() => {
-    getOrganisasi();
-  }, []);
-
+    getOrganisasi(currentPage);
+  }, [currentPage]);
   return (
     <div className="flex">
       <Sidebar />
@@ -112,7 +174,8 @@ function Organisasi() {
                 </h1>
                 <button
                   onClick={openModal}
-                  className="inline-block rounded bg-[#776b5d] px-4 py-2 text-sm font-medium text-white transition hover:scale-110 hover:shadow-xl focus:outline-none focus:ring active:bg-[#776b5d] ml-0 sm:ml-4">
+                  className="inline-block rounded bg-[#776b5d] px-4 py-2 text-sm font-medium text-white transition hover:scale-110 hover:shadow-xl focus:outline-none focus:ring active:bg-[#776b5d] ml-0 sm:ml-4"
+                >
                   Tambah
                 </button>
               </div>
@@ -132,7 +195,8 @@ function Organisasi() {
                       <svg
                         className="fill-current h-4 w-4"
                         xmlns="http://www.w3.org/2000/svg"
-                        viewBox="0 0 20 20">
+                        viewBox="0 0 20 20"
+                      >
                         <path d="M9 11l3-3 3 3m-3 3v-6" />
                       </svg>
                     </div>
@@ -160,6 +224,12 @@ function Organisasi() {
                     <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-800">
                       Nama Organisasi
                     </th>
+                    <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-800">
+                      Create At
+                    </th>
+                    <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-800">
+                      Update At
+                    </th>
                     <th className="whitespace-nowrap px-4 py-2 font-bold text-gray-800 text-center">
                       Aksi
                     </th>
@@ -175,17 +245,26 @@ function Organisasi() {
                         <td className="whitespace-nowrap px-4 py-2 text-gray-700">
                           {organisasi.nama_organisasi}
                         </td>
+                        <td className="whitespace-nowrap px-4 py-2 text-gray-700">
+                          {organisasi.createdDate}
+                        </td>
+                        <td className="whitespace-nowrap px-4 py-2 text-gray-700">
+                          {organisasi.updatedDate}
+                        </td>
+
                         <td className="whitespace-nowrap flex justify-center gap-3 px-4 py-2 text-gray-700">
                           <Link
                             onClick={() => openModalEdit(organisasi.id)}
                             className="block rounded-md bg-blue-400 border border-transparent fill-white p-2 text-sm font-medium text-white transition-all duration-200 hover:shadow-md hover:bg-transparent hover:fill-blue-400 hover:border-blue-400"
-                            title="Edit">
+                            title="Edit"
+                          >
                             <svg
                               xmlns="http://www.w3.org/2000/svg"
                               id="Outline"
                               viewBox="0 0 24 24"
                               width="18"
-                              height="18">
+                              height="18"
+                            >
                               <path d="M18.656.93,6.464,13.122A4.966,4.966,0,0,0,5,16.657V18a1,1,0,0,0,1,1H7.343a4.966,4.966,0,0,0,3.535-1.464L23.07,5.344a3.125,3.125,0,0,0,0-4.414A3.194,3.194,0,0,0,18.656.93Zm3,3L9.464,16.122A3.02,3.02,0,0,1,7.343,17H7v-.343a3.02,3.02,0,0,1,.878-2.121L20.07,2.344a1.148,1.148,0,0,1,1.586,0A1.123,1.123,0,0,1,21.656,3.93Z" />
                               <path d="M23,8.979a1,1,0,0,0-1,1V15H18a3,3,0,0,0-3,3v4H5a3,3,0,0,1-3-3V5A3,3,0,0,1,5,2h9.042a1,1,0,0,0,0-2H5A5.006,5.006,0,0,0,0,5V19a5.006,5.006,0,0,0,5,5H16.343a4.968,4.968,0,0,0,3.536-1.464l2.656-2.658A4.968,4.968,0,0,0,24,16.343V9.979A1,1,0,0,0,23,8.979ZM18.465,21.122a2.975,2.975,0,0,1-1.465.8V18a1,1,0,0,1,1-1h3.925a3.016,3.016,0,0,1-.8,1.464Z" />
                             </svg>
@@ -193,14 +272,16 @@ function Organisasi() {
                           <Link
                             onClick={() => deleteOrganisasi(organisasi.id)}
                             className="block rounded-md bg-red-500 border border-transparent fill-white p-2 text-sm font-medium text-white transition-all duration-200 hover:shadow-md hover:bg-transparent hover:fill-red-500 hover:border-red-500"
-                            title="Hapus">
+                            title="Hapus"
+                          >
                             <svg
                               xmlns="http://www.w3.org/2000/svg"
                               width="18"
                               height="18"
                               fill="inherit"
                               className="bi bi-trash"
-                              viewBox="0 0 16 16">
+                              viewBox="0 0 16 16"
+                            >
                               <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z" />
                               <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z" />
                             </svg>
@@ -212,17 +293,26 @@ function Organisasi() {
                 </tbody>
               </table>
             </div>
+            {/* Pagination */}
             <ol className="flex justify-center gap-1 text-xs font-medium">
               <li>
+                {/* Menangani halaman sebelumnya */}
                 <a
                   href="#"
-                  className="inline-flex size-8 items-center justify-center rounded border border-gray-500 bg-white text-gray-900 rtl:rotate-180">
+                  onClick={() => handlePageChange(currentPage - 1)}
+                  className={`inline-flex size-8 items-center justify-center rounded border border-gray-500 bg-white text-gray-900 rtl:rotate-180 ${
+                    currentPage === 1
+                      ? "opacity-50 cursor-not-allowed bg-gray-200"
+                      : ""
+                  }`}
+                >
                   <span className="sr-only">Prev Page</span>
                   <svg
                     xmlns="http://www.w3.org/2000/svg"
                     className="h-3 w-3"
                     viewBox="0 0 20 20"
-                    fill="currentColor">
+                    fill="currentColor"
+                  >
                     <path
                       fillRule="evenodd"
                       d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
@@ -232,44 +322,40 @@ function Organisasi() {
                 </a>
               </li>
 
+              {/* Menampilkan nomor halaman */}
+              {Array.from({ length: pages }, (_, i) => (
+                <li key={i}>
+                  <a
+                    href="#"
+                    onClick={() => handlePageChange(i)} // Tidak perlu menambahkan 1 karena kita sudah mulai dari 0
+                    className={`block size-8 rounded border border-gray-500 bg-white text-center leading-8 text-gray-900 ${
+                      i === currentPage ? "bg-gray-500 text-white" : ""
+                    }`}
+                  >
+                    {i + 1}{" "}
+                    {/* Tambahkan 1 untuk menampilkan nomor halaman yang dimulai dari 1 */}
+                  </a>
+                </li>
+              ))}
+
+              {/* Menangani halaman berikutnya */}
               <li>
                 <a
                   href="#"
-                  className="block size-8 rounded border border-gray-500 bg-white text-center leading-8 text-gray-900">
-                  1
-                </a>
-              </li>
-
-              <li className="block size-8 rounded border border-gray-500 bg-white text-center leading-8 text-gray-900">
-                2
-              </li>
-
-              <li>
-                <a
-                  href="#"
-                  className="block size-8 rounded border border-gray-500 bg-white text-center leading-8 text-gray-900">
-                  3
-                </a>
-              </li>
-
-              <li>
-                <a
-                  href="#"
-                  className="block size-8 rounded border border-gray-500 bg-white text-center leading-8 text-gray-900">
-                  4
-                </a>
-              </li>
-
-              <li>
-                <a
-                  href="#"
-                  className="inline-flex size-8 items-center justify-center rounded border border-gray-500 bg-white text-gray-900 rtl:rotate-180">
+                  onClick={() => handlePageChange(currentPage + 1)}
+                  className={`inline-flex size-8 items-center justify-center rounded border border-gray-500 bg-white text-gray-900 ${
+                    currentPage >= pages
+                      ? "opacity-50 cursor-not-allowed bg-gray-200"
+                      : ""
+                  }`}
+                >
                   <span className="sr-only">Next Page</span>
                   <svg
                     xmlns="http://www.w3.org/2000/svg"
                     className="h-3 w-3"
                     viewBox="0 0 20 20"
-                    fill="currentColor">
+                    fill="currentColor"
+                  >
                     <path
                       fillRule="evenodd"
                       d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
@@ -300,7 +386,8 @@ function Organisasi() {
                         <div className="mt-4 mb-4 sm:w-full">
                           <label
                             htmlFor="organisasi"
-                            className="relative block overflow-hidden border-b border-gray-400 bg-transparent pt-3 focus-within:border-[#776b5d]">
+                            className="relative block overflow-hidden border-b border-gray-400 bg-transparent pt-3 focus-within:border-[#776b5d]"
+                          >
                             <input
                               className="peer h-8 w-full border-none bg-transparent p-0 placeholder-transparent focus:border-transparent focus:outline-none focus:ring-0 sm:text-sm"
                               type="text"
@@ -323,12 +410,14 @@ function Organisasi() {
                     <div className="flex flex-col sm:flex-row justify-between">
                       <button
                         onClick={closeModalEdit}
-                        className="w-full sm:w-auto rounded-md border border-red-500 bg-red-500 px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-red-500 focus:outline-none active:text-white active:bg-red-400">
+                        className="w-full sm:w-auto rounded-md border border-red-500 bg-red-500 px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-red-500 focus:outline-none active:text-white active:bg-red-400"
+                      >
                         Kembali
                       </button>
                       <button
                         onClick={putOrganisasi}
-                        className="w-full sm:w-auto mt-4 sm:mt-0 rounded-md border border-[#776B5D] bg-[#776B5D] px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-[#776B5D] focus:outline-none active:text-white active:bg-[#776d5b]">
+                        className="w-full sm:w-auto mt-4 sm:mt-0 rounded-md border border-[#776B5D] bg-[#776B5D] px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-[#776B5D] focus:outline-none active:text-white active:bg-[#776d5b]"
+                      >
                         Simpan
                       </button>
                     </div>
@@ -357,7 +446,8 @@ function Organisasi() {
                         <div className="mt-4 mb-4 sm:w-full">
                           <label
                             htmlFor="organisasi"
-                            className="relative block overflow-hidden border-b border-gray-400 bg-transparent pt-3 focus-within:border-[#776b5d]">
+                            className="relative block overflow-hidden border-b border-gray-400 bg-transparent pt-3 focus-within:border-[#776b5d]"
+                          >
                             <input
                               className="peer h-8 w-full border-none bg-transparent p-0 placeholder-transparent focus:border-transparent focus:outline-none focus:ring-0 sm:text-sm"
                               type="text"
@@ -380,12 +470,14 @@ function Organisasi() {
                     <div className="flex flex-col sm:flex-row justify-between">
                       <button
                         onClick={closeModal}
-                        className="w-full sm:w-auto rounded-md border border-red-500 bg-red-500 px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-red-500 focus:outline-none active:text-white active:bg-red-400">
+                        className="w-full sm:w-auto rounded-md border border-red-500 bg-red-500 px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-red-500 focus:outline-none active:text-white active:bg-red-400"
+                      >
                         Kembali
                       </button>
                       <button
                         onClick={addOrganisasi}
-                        className="w-full sm:w-auto mt-4 sm:mt-0 rounded-md border border-[#776B5D] bg-[#776B5D] px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-[#776B5D] focus:outline-none active:text-white active:bg-[#776d5b]">
+                        className="w-full sm:w-auto mt-4 sm:mt-0 rounded-md border border-[#776B5D] bg-[#776B5D] px-6 py-2 text-xs font-medium text-white transition hover:bg-transparent hover:text-[#776B5D] focus:outline-none active:text-white active:bg-[#776d5b]"
+                      >
                         Simpan
                       </button>
                     </div>
